requirements.txt
Flask==2.3.2
pymongo==4.4.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
python-dateutil==2.8.2
uuid==1.30